Revert 'skip_existing' on the chart releaser
This correctly fixes the "trying to make a release that already exists"
issue, but the action does not honor that when it comes to the update of
the index.yaml file for the github-pages. So we end up with multiple
entries for the same chart version, with the same link to the first
artifacts but different, invalid hashes.
See https://github.com/helm/chart-releaser-action/issues/209#issuecomment-3051790860

For now, it is preferable to have the workflow failure instead of
generating garbage in the index.yaml file