Manually trigger and run the chart releaser
The chart-releaser-action behavior (and options) does not fit well our
wanted workflow for this repo. It wants to release all modified charts,
even when their version did not change which leads to errors.
On our side we would like to be able to do development commits on the
main branch without having to release a chart and control when and which
chart to release.

This can be done, with a bit more work, by not using the action but the
chart releaser CLI directly to package and release a single selected
chart on a manual trigger.