Use helm comment instead of YAML comment for copyrights
This should avoid issues with '{{- if' blocks where the first YAML line
of a ressource might get merged with the last comment line
This then causes issues because we are missing a data line in the
kubernetes resource definition.
It's easier to use Helm comments to avoid injecting them into the YAML
than making sure the rendered YAML is valid in all cases.